requirements-build.txt
# Build dependencies for creating Windows executable
pyinstaller>=6.0.0
pillow>=10.0.0
